# api/openapi.yaml
openapi: 3.1.0
info:
  title: Halo management API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The JSON API behind the Halo console and account portal. Anything the web interface does, a script can do with the same requests.

    **Authentication.** Send either the `halo_session` cookie of a signed-in person, or an API key of a service account as `Authorization: Bearer hlk_…`. API keys work only when they have the `api` scope, act with the roles of their service account, and are ignored on the personal routes under `/me` and on the sign-in routes under `/auth`. The personal routes under `/me/` (with a trailing segment, not `/me` itself) also accept an access token that `halo login` obtained for the built-in Halo CLI application.

    **Roles.** Each operation states the roles it requires. A global administrator passes every role check. "Any administrator" means security, user, helpdesk or application administrator, or auditor.

    **Errors.** Every error is `{"error": {"code": "ERR_…", "message": "…"}}`. Match on `code`; show `message` to people. Besides the codes listed per operation, any request can fail with `500` and `ERR_INTERNAL`, and unknown routes answer `404` with `ERR_NO_ROUTE`. Requests with a method other than GET, HEAD or OPTIONS that come from another website get `403` with `ERR_CROSS_SITE`.

...
servers:
  - url: /api/v1
security:
  - session: []
  - apiKey: []
...
paths:
...
  /users/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: getUser
      summary: Get one person
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Change a person's profile
      description: 'Roles: user or helpdesk administrator. Helpdesk administrators can change title, department and location only. Only a global administrator can change a person who holds an administrator role. Send only the fields to change; an empty `managerId` removes the manager.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string}
      responses:
        '200':
          description: The updated person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
...
  /applications:
    get:
      tags: [Applications]
      operationId: listApplications
      summary: List every application
      description: 'Roles: any administrator. Includes credentials without secret values, assigned groups and sign-in statistics.'
      responses:
        '200':
          description: The applications.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Application'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Applications]
      operationId: createApplication
      summary: Register an application
      description: 'Roles: application administrator. OpenID Connect web applications and services get a client secret, returned once. SAML applications take their settings in `saml` and are always `web`.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, protocol, type]
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                protocol: {type: string, enum: [oidc, oauth, saml]}
                type: {type: string, enum: [web, spa, native, service]}
                homepage: {type: string, format: uri}
                redirectUris:
                  type: array
                  description: Required for OpenID Connect applications other than services. https only, except http://localhost and http://127.0.0.1.
                  items: {type: string, format: uri}
                postLogoutUris:
                  type: array
                  items: {type: string, format: uri}
                scopes:
                  type: array
                  items: {type: string}
                groupIds:
                  type: array
                  items: {type: string}
                setupGuide: {type: string, enum: [grafana, forgejo, generic-oidc, kubernetes, generic-saml, aws-iam-identity-center, slack]}
                saml: {$ref: '#/components/schemas/SAMLInput'}
      responses:
        '201':
          description: The application and, for web applications and services, the client secret.
          content:
            application/json:
              schema:
                type: object
                required: [application, clientSecret]
                properties:
                  application: {$ref: '#/components/schemas/Application'}
                  clientSecret: {type: [string, 'null'], examples: [hls_…]}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /applications/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Applications]
      operationId: getApplication
      summary: Get one application
      description: 'Roles: any administrator. SAML applications include a `saml` object.'
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Applications]
      operationId: updateApplication
      summary: Change an application
      description: 'Roles: application administrator. Send only the fields to change. SAML applications accept only name, description, homepage and owner. OpenID Connect applications must keep the `openid` scope.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                homepage: {type: string}
                redirectUris:
                  type: array
                  items: {type: string}
                postLogoutUris:
                  type: array
                  items: {type: string}
                scopes:
                  type: array
                  items: {type: string}
                tokenPolicy:
                  type: object
                  properties:
                    accessTokenTtl: {type: integer, minimum: 60, maximum: 86400}
                    idTokenTtl: {type: integer, minimum: 60, maximum: 86400, description: Services may use 0.}
                    refreshTokenTtl: {type: integer, minimum: 0, maximum: 7776000, description: 0 stops issuing refresh tokens.}
                    rotation: {type: boolean}
                owner: {type: string, description: Id of a person. An empty string removes the owner.}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
    delete:
      tags: [Applications]
      operationId: deleteApplication
      summary: Delete an application with its credentials and group assignments
      description: 'Roles: application administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: '`ERR_BUILT_IN`: the Halo CLI application cannot be deleted; disable it instead.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
...
  /applications/{id}/groups:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Applications]
      operationId: setApplicationGroups
      summary: Replace the groups assigned to an application
      description: 'Roles: application administrator. Send the complete list; an empty list removes every assignment.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupIds]
              properties:
                groupIds:
                  type: array
                  items: {type: string}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /applications/{id}/saml:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Applications]
      operationId: setApplicationSAML
      summary: Change a SAML application's settings
      description: 'Roles: application administrator. Send any of the fields; `metadataXml` imports the entity ID and ACS URLs from the service provider''s metadata.'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SAMLInput'}
      responses:
        '200': {$ref: '#/components/responses/ApplicationResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: halo_session
      description: The session cookie Halo sets when a person signs in.
    apiKey:
      type: http
      scheme: bearer
      description: An API key of a service account, starting with `hlk_`, with the `api` scope. Requests act with the service account's roles.
    cliAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token that `halo login` obtained for the built-in Halo CLI application. Accepted only on routes under `/me/`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string}
...
  responses:
    NoContent:
      description: Done. No body.
    BadJSON:
      description: '`ERR_INVALID_JSON`: the body is not valid JSON, is larger than 1 MiB, or has an unknown field.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthenticated:
      description: '`ERR_UNAUTHENTICATED`: no valid session or key.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: '`ERR_FORBIDDEN`: your roles do not allow this. `ERR_CROSS_SITE` on changes from another website.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: '`ERR_NOT_FOUND`: the item does not exist.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: '`ERR_CONFLICT`: the change conflicts with the current state, such as a name that is already used.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Invalid:
      description: '`ERR_INVALID`: a field is missing or invalid. The message names it.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...